Unified filterorderbyoffsetlimit tests
